Grid size and pre-filtering added.
This is now the best version I have, and I'm still improving it. I made a silly error checking for orange on red blocks in the first cycle in previous versions; you should start with yellow on orange blocks because that's the first cycle where blocks can be placed. I added a pre-filter check to test all candidate vertices to see if you can place a cube of size =1, and if not, to eliminate this from the candidate list. This greatly speeds things up by eliminating lots of options quickly and preventing them from being reconsidered in subsequent cycles. I've also implemented a grid feature that looks for collisions between the candidate cube and existing cubes by searching the local area, broken into grid-size portions. Mostly just breaking things up into 1x1x1 sections is the fastest, but at N=14 and above, it starts to be faster to search 2x2x2 sections. It might be faster at higher grid sizes for higher N values, but I haven't tested above n=17 yet, and so far 2 is the highest value that can give a potential calculation speed boost.